+++ recipes/linuxbridge.rb
@@ -35,4 +35,5 @@ service 'neutron-plugin-linuxbridge-agent' do
   service_name platform_options['neutron_linuxbridge_agent_service']
   supports status: true, restart: true
   action :enable
+  subscribes :restart, 'template[/etc/neutron/neutron.conf]'
 end